services/user/requirements.dev.txt
annotated-types==0.6.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
pydantic==2.4.2
pydantic_core==2.10.1
python-decouple==3.8
typing_extensions==4.8.0
